moved onchildchanged listener into custom hook
